pam_lastlog: fix format string
gcc -Wformat justly complains:
format '%d' expects argument of type 'int', but argument 5 has type 'time_t'

* modules/pam_lastlog/pam_lastlog.c (pam_sm_authenticate): Fix format
string.

+++ modules/pam_lastlog/pam_lastlog.c
@@ -628,7 +628,8 @@ pam_sm_authenticate(pam_handle_t *pamh, int flags,
     lltime = (time(NULL) - lltime) / (24*60*60);
 
     if (lltime > inactive_days) {
-        pam_syslog(pamh, LOG_INFO, "user %s inactive for %d days - denied", user, lltime);
+        pam_syslog(pamh, LOG_INFO, "user %s inactive for %ld days - denied",
+		   user, (long) lltime);
         return PAM_AUTH_ERR;
     }
